5ZCMF7C name the shared api modules by concept: issue-projection, issue-targets; contributor name resolution goes home
